Refactor transaction assets
