+++ packages/bierzo-wallet/src/logic/transactions/types/BwSendTransaction/ui/SendTxHeader/index.tsx
@@ -9,8 +9,8 @@ import { history } from "../../../../../../routes";
 import { PAYMENT_ROUTE } from "../../../../../../routes/paths";
 import { ProcessedSendTransaction } from "../../../../../../store/notifications";
 import { itemBackground } from "../../../../../../theme/css";
+import sendTx from "../../../../assets/transactionSend.svg";
 import receiveTx from "../assets/transactionReceive.svg";
-import sendTx from "../assets/transactionSend.svg";
 import Msg from "./MsgSuccess";
 
 interface ItemProps {
